refactor: переименовал файлы для вебсокета, ws/routes - для роутов вс, ws/manager - вс менеджер
